Update app : app2의 내용으로 대체
